[FEAT] 워크스페이스 세부 내용 조회에 total 필드 추가

+++ module-monitor/src/main/java/com/xiilab/modulemonitor/repository/K8sMonitorRepositoryImpl.java
@@ -281,9 +281,9 @@ public class K8sMonitorRepositoryImpl implements K8sMonitorRepository {
 		try (KubernetesClient kubernetesClient = monitorK8SAdapter.configServer()) {
 			List<Node> nodeList = kubernetesClient.nodes().list().getItems();
 			List<Pod> podList = kubernetesClient.pods().list().getItems();
-			int cpu = totalCapacity(nodeList, "CPU");
-			int mem = totalCapacity(nodeList, "MEM");
-			int gpu = totalCapacity(nodeList, "GPU");
+			int cpu = (int)totalCapacity(nodeList, "CPU");
+			int mem = (int)totalCapacity(nodeList, "MEM");
+			int gpu = (int)totalCapacity(nodeList, "GPU");
 			String totalCpuRequests = totalRequests(nodeList, podList, "CPU");
 			String totalMemRequests = totalRequests(nodeList, podList, "MEM");
 			String totalGpuRequests = totalRequests(nodeList, podList, "GPU");
